+++ components/loconetmessages/CMakeLists.txt
@@ -3,7 +3,7 @@ set(COMPONENT_SRCS
 			"src/loconet/message/message.cpp"
 			"src/loconet/message/busy.cpp"
 			"src/loconet/message/inputreport.cpp"
-			"src/loconet/message/messageparser.cpp"
+			"src/loconet/message/parser.cpp"
 			"src/loconet/receiver.cpp"
 			"src/loconet/dispatcher.cpp"
 )
